AdditionFloat

+++ SubtractionFloat/SubtractionFloat.java
@@ -1,8 +1,8 @@
-public class SubtractionFloat{
-  public static void main(String args[]){
-    double a,b,c;
-    a = 1.4;
-    b = 3.5;
+public class SubtractionFloat {
+  public static void main(String args[]) {
+    float a,b,c;
+    a = (float)1.4;
+    b = (float)3.5;
     c = a-b;
     System.out.println("Float Difference = " + c);
   }
